move stories and markdown files to core if the component lives in core
