+++ ui/lib/core/stories/icon.stories.js
@@ -2,11 +2,11 @@
 import hbs from 'htmlbars-inline-precompile';
 import { storiesOf } from '@storybook/ember';
 import notes from './icon.md';
-import icons from '../node_modules/@hashicorp/structure-icons/dist/index.js';
+import icons from '../../../node_modules/@hashicorp/structure-icons/dist/index.js';
 import { withKnobs, select } from '@storybook/addon-knobs';
 
 storiesOf('Icon/', module)
-  .addParameters({ options: { showPanel: true} })
+  .addParameters({ options: { showPanel: true } })
   .addDecorator(withKnobs())
   .add(
     'Icon',
