+++ ui/lib/core/stories/info-table-row.md
