+++ ui/lib/core/stories/layout-loading.stories.js
@@ -5,12 +5,14 @@ import notes from './layout-loading.md';
 
 storiesOf('Loading/LayoutLoading/', module)
   .addParameters({ options: { showPanel: true } })
-  .add(`LayoutLoading`, () => ({
-    template: hbs`
+  .add(
+    `LayoutLoading`,
+    () => ({
+      template: hbs`
         <h5 class="title is-5">Layout Loading</h5>
         <LayoutLoading/>
     `,
-    context: {},
-  }),
-  {notes}
-);
+      context: {},
+    }),
+    { notes }
+  );
