+++ ui/lib/core/stories/masked-input.md
