+++ ui/lib/core/stories/message-error.stories.js
@@ -5,21 +5,23 @@ import notes from './message-error.md';
 
 let model = Ember.Object.create({
   adapterError: {
-    message: 'This is an adapterError on the model'
+    message: 'This is an adapterError on the model',
   },
-  isError: true
+  isError: true,
 });
 
 storiesOf('MessageError/', module)
   .addParameters({ options: { showPanel: true } })
-  .add(`MessageError`, () => ({
-    template: hbs`
+  .add(
+    `MessageError`,
+    () => ({
+      template: hbs`
       <h5 class="title is-5">Message Error</h5>
       <MessageError @model={{model}} />
     `,
-    context: {
-      model
-    }
-  }),
-  {notes}
-);
+      context: {
+        model,
+      },
+    }),
+    { notes }
+  );
