+++ ui/lib/core/stories/popup-menu.stories.js
@@ -3,11 +3,12 @@ import hbs from 'htmlbars-inline-precompile';
 import { storiesOf } from '@storybook/ember';
 import notes from './popup-menu.md';
 
-
 storiesOf('PopupMenu/', module)
   .addParameters({ options: { showPanel: true } })
-  .add(`PopupMenu`, () => ({
-    template: hbs`
+  .add(
+    `PopupMenu`,
+    () => ({
+      template: hbs`
         <h5 class="title is-5">Popup Menu</h5>
         <PopupMenu>
           <nav class="menu">
@@ -21,7 +22,7 @@ storiesOf('PopupMenu/', module)
           </nav>
         </PopupMenu>
     `,
-    context: {},
-  }),
-  {notes}
-);
+      context: {},
+    }),
+    { notes }
+  );
