+++ ui/lib/core/stories/toolbar/toolbar-link.stories.js
@@ -2,35 +2,33 @@
 import hbs from 'htmlbars-inline-precompile';
 import { storiesOf } from '@storybook/ember';
 import { withKnobs, select, text } from '@storybook/addon-knobs';
-import notes from './toolbar-secret-link.md';
-
+import notes from './toolbar-link.md';
 
 storiesOf('Toolbar/', module)
   .addParameters({ options: { showPanel: true } })
   .addDecorator(withKnobs())
-  .add(`ToolbarSecretLink`,() => ({
-    template: hbs`
+  .add(
+    `ToolbarLink`,
+    () => ({
+      template: hbs`
       <h5 class="title is-5">ToolbarLink</h5>
       <div style="width: 400px;">
         <Toolbar>
           <ToolbarActions>
-            <ToolbarSecretLink
-              @secret={{model.id}}
-              @mode="edit"
-              @data-test-edit-link=true
-              @replace=true
+            <ToolbarLink
+              @params={{array '#'}}
               @type={{type}}
             >
               {{label}}
-            </ToolbarSecretLink>
+            </ToolbarLink>
           </ToolbarActions>
         </Toolbar>
       </div>
     `,
-    context: {
-      type: select('Type', ['', 'add']),
-      label: text('Button text', 'Edit role'),
-    },
-  }),
-  {notes}
-);
+      context: {
+        type: select('Type', ['', 'add']),
+        label: text('Button text', 'Edit secret'),
+      },
+    }),
+    { notes }
+  );
