+++ ui/lib/core/stories/toolbar/toolbar-secret-link.stories.js
@@ -2,32 +2,36 @@
 import hbs from 'htmlbars-inline-precompile';
 import { storiesOf } from '@storybook/ember';
 import { withKnobs, select, text } from '@storybook/addon-knobs';
-import notes from './toolbar-link.md';
-
+import notes from './toolbar-secret-link.md';
 
 storiesOf('Toolbar/', module)
   .addParameters({ options: { showPanel: true } })
   .addDecorator(withKnobs())
-  .add(`ToolbarLink`,() => ({
-    template: hbs`
+  .add(
+    `ToolbarSecretLink`,
+    () => ({
+      template: hbs`
       <h5 class="title is-5">ToolbarLink</h5>
       <div style="width: 400px;">
         <Toolbar>
           <ToolbarActions>
-            <ToolbarLink
-              @params={{array '#'}}
+            <ToolbarSecretLink
+              @secret={{model.id}}
+              @mode="edit"
+              @data-test-edit-link=true
+              @replace=true
               @type={{type}}
             >
               {{label}}
-            </ToolbarLink>
+            </ToolbarSecretLink>
           </ToolbarActions>
         </Toolbar>
       </div>
     `,
-    context: {
-      type: select('Type', ['', 'add']),
-      label: text('Button text', 'Edit secret'),
-    },
-  }),
-  {notes}
-);
+      context: {
+        type: select('Type', ['', 'add']),
+        label: text('Button text', 'Edit role'),
+      },
+    }),
+    { notes }
+  );
