+++ ui/lib/core/stories/toolbar/toolbar.stories.js
@@ -7,8 +7,10 @@ import notes from './toolbar.md';
 storiesOf('Toolbar/', module)
   .addParameters({ options: { showPanel: true } })
   .addDecorator(withKnobs())
-  .add(`Toolbar`, () => ({
-    template: hbs`
+  .add(
+    `Toolbar`,
+    () => ({
+      template: hbs`
       <h5 class="title is-5">Toolbar</h5>
       <section class="box is-fullwidth is-shadowless">
         <h5 class="title is-6">Example for list views</h5>
@@ -72,9 +74,9 @@ storiesOf('Toolbar/', module)
         </Toolbar>
       </section>
     `,
-    context: {
-      example: select('Example', ['List', 'Show', 'Code editor'], 'List'),
-    },
-  }),
-  {notes}
-);
+      context: {
+        example: select('Example', ['List', 'Show', 'Code editor'], 'List'),
+      },
+    }),
+    { notes }
+  );
