+++ ui/lib/core/stories/vault-logo-spinner.md
